+++ main/ui.c
@@ -59,9 +59,9 @@ static void tileDeleteCb(lv_event_t *event) {
  * Process UI state changes
  */
 
-static void uiEventHandler(void *handler_arg, esp_event_base_t base, long newState, void *event_data) {
+static void uiEventHandler(void *handler_arg, esp_event_base_t base, int newState, void *event_data) {
     if (newState >= ARRAY_SIZE(uiScreens)) {
-        ESP_LOGI(TAG, "Request for unknown UI state %ld", newState);
+        ESP_LOGI(TAG, "Request for unknown UI state %d", (int)newState);
         return;
     }
     currentScreen = uiScreens[newState];
@@ -96,7 +96,8 @@ static void uiEventHandler(void *handler_arg, esp_event_base_t base, long newSta
 }
 
 void showStats() {
-    printf("heap free, = %ld smallest  = %ld\n", xPortGetFreeHeapSize(), xPortGetMinimumEverFreeHeapSize());
+    printf("heap free, = %ld smallest  = %ld\n",
+           (unsigned long)xPortGetFreeHeapSize(), (unsigned long)xPortGetMinimumEverFreeHeapSize());
     /*
     TaskStatus_t tasks[30];
     uint32_t runTime;
